+++ tests/services/test_api_test.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2020, 2021 Roald Nefs <info@roaldnefs.com>
+# Copyright (C) 2021 Roald Nefs <info@roaldnefs.com>
 #
 # This file is part of python-transip.
 #
@@ -17,39 +17,29 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with python-transip.  If not, see <https://www.gnu.org/licenses/>.
 
-from typing import List, Tuple, Any, Dict
 import responses  # type: ignore
 import unittest
 
 from transip import TransIP
-from transip.v6.objects import Invoice
+
 from tests.utils import load_responses_fixtures
 
-class InvoicesTest(unittest.TestCase):
-    """Test the InvoiceService."""
+
+class ApiTestTest(unittest.TestCase):
+    """Test the ApiTestService."""
 
     client: TransIP
 
     @classmethod
     def setUpClass(cls) -> None:
-        """Set up a minimal TransIP client for using the invoice services."""
+        """Set up a minimal TransIP client for using the API test services."""
         cls.client = TransIP(access_token='ACCESS_TOKEN')
 
     def setUp(self) -> None:
-        """Setup mocked responses for the '/invoices' endpoint."""
-        load_responses_fixtures("account.json")
+        """Setup mocked responses for the '/api-test' endpoint."""
+        load_responses_fixtures("general.json")
 
     @responses.activate
-    def test_list(self) -> None:
-        invoices: List[Invoice] = self.client.invoices.list()  # type: ignore
-        invoice: Invoice = invoices[0]
-    
-        assert len(invoices) == 1
-        assert invoice.get_id() == "F0000.1911.0000.0004"  # type: ignore
-
-    @responses.activate
-    def test_get(self) -> None:
-        invoice_id: str = "F0000.1911.0000.0004"
-        invoice: Invoice = self.client.invoices.get(invoice_id)  # type: ignore
-
-        assert invoice.get_id() == "F0000.1911.0000.0004"  # type: ignore
+    def test_test(self) -> None:
+        """Test if the API test returns True."""
+        self.assertTrue(self.client.api_test.test())  # type: ignore