+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes in **python-transip** are documented below.
 
 ## [Unreleased]
 ### Added
-- This CHANGELOG file to be able to list all notable changes for each version of **python-transip**.
+- This `CHANGELOG.md` file to be able to list all notable changes for each version of **python-transip**.
+- The `transip.v6.objects.ApiTestService` service to allow calling the test resource to make sure everything is working.
 
 [Unreleased]: https://github.com/roaldnefs/python-transip/compare/v0.3.0...HEAD